Change set_state in sensor to float fixed up recv data

+++ esphome/components/nextion/nextion_sensor.h
@@ -20,7 +20,7 @@ class NextionSensor : public NextionComponent, public sensor::Sensor, public Pol
   void set_precision(uint8_t precision) { this->precision_ = precision; }
 
   void process_sensor(char *variable_name, int state);
-  void set_state(int state);
+  void set_state(float state);
 
  protected:
   uint8_t precision_ = 0;
